add new task

+++ 0x0B-python-input_output/5-save_to_json_file.py
@@ -2,6 +2,7 @@
 """module for  Python - Input/Output
     """
 import json
+import os
 
 
 def save_to_json_file(my_obj, filename):
@@ -11,5 +12,8 @@ def save_to_json_file(my_obj, filename):
         my_obj (string): to be converted
         filename: to put into
     """
+    if os.path.exists(filename):
+        with open(filename, "a", encoding="utf-8") as f_new:
+            json.dump(my_obj, f_new)
     with open(filename, "w", encoding="utf-8") as f_new:
-        json.dump(my_obj, f_new)
+            json.dump(my_obj, f_new)
